requirements.txt
flask
openai
pinecone
ijson
sentence_transformers
tqdm
joblib
wordninja
farm-haystack[inference]